requirements.txt
torch
numpy
matplotlib
hydra-core
omegaconf
tqdm
opencv-python